Set default list factory (depends on the platform)
