+++ impl/default/common/build.gradle
@@ -1,0 +1,5 @@
+apply from: rootProject.file('gradle/common.gradle')
+
+dependencies {
+    compile project(':collekt-api-common')
+}
