+++ impl/default/js/src/main/kotlin/com/github/jcornaz/collekt/DefaultFactories.kt
@@ -1,0 +1,6 @@
+package com.github.jcornaz.collekt
+
+import com.github.jcornaz.collekt.api.PersistentListFactory
+import com.github.jcornaz.collekt.stdlib.StdlibList
+
+public actual var DefaultListFactory: PersistentListFactory = StdlibList
