+++ impl/default/jvm6/src/main/kotlin/com/github/jcornaz/collekt/DefaultFactories.kt
@@ -1,0 +1,6 @@
+package com.github.jcornaz.collekt
+
+import com.github.jcornaz.collekt.api.PersistentListFactory
+import com.github.jcornaz.collekt.dexx.DexxVector
+
+public actual var DefaultListFactory: PersistentListFactory = DexxVector
