+++ settings.gradle
@@ -13,6 +13,7 @@ include 'collekt-dexx'
 include 'collekt-vavr'
 include 'collekt-pcollections'
 
+include 'collekt-default-common'
 include 'collekt-default-jvm6'
 include 'collekt-default-jvm8'
 include 'collekt-default-js'
@@ -36,6 +37,7 @@ project(':collekt-dexx').projectDir = file('impl/dexx')
 project(':collekt-vavr').projectDir = file('impl/vavr')
 project(':collekt-pcollections').projectDir = file('impl/pcollections')
 
+project(':collekt-default-common').projectDir = file('impl/default/common')
 project(':collekt-default-jvm6').projectDir = file('impl/default/jvm6')
 project(':collekt-default-jvm8').projectDir = file('impl/default/jvm8')
 project(':collekt-default-js').projectDir = file('impl/default/js')
